+++ sdk/eventhub/event-hubs/CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Breaking Changes
 - `MessageWithMetadata` has been renamed to `MessageContent`.
 - `MessageContent`'s `body` has been renamed to `data`.
+- `MessageAdapter`'s `consumeMessage` and `produceMessage` have been renamed to `consume` and `produce`.
 
 ### Bugs Fixed